Change return type of `determine_os()` to `OS`
This commit changes the value returned by `determine_os()`
from an `OperatingSystem` dataclass to the `OS` enum (of
`OperatingSystem` dataclasses). This should make it possible
to do exhaustive pattern matching on the result of
`determine_os()`.

To do this, we had to make the `OS` class inherit from `Enum`
instead of `EnumType`. To make this work, we had to stop using
`OperatingSystemMeta` as the metaclass for `OperatingSystem`
(there were inheritance conflict otherwise). So, we moved
`OperatingSystemMeta`'s functionality directly into
`OperatingSystem`, which then meant we could get rid of
`OperatingSystemMeta` entirely.

We also had to add a manually-defined `__init__` method to
`OperatingSystem` since we wanted the option to keep calling
`OperatingSystem(..., unix_like=...)` when declaring the
values of the `OS` Enum members.

This commit changes our public API, so we should bump the
version number to v2.0.

Changes to be committed:
	modified:   src/lmn/cx/y22/operating_systems/_operating_systems.py