Fix failing tests (requires API break)
The `xdg` value for the `OperatingSystem` instances was supposed
to have a default value that depended on that instance's `unix_like`
value, but that wasn't working.

After playing around with metaclasses and decorators and things,
I decided the easiest way forward was just to handle it as a
`__post_init__` method and call it a day. No need to think about
`__call__`, etc., to modify arguments being passed to
`OperatingSystem.__init__`.

As a side effect of all of this, `OperatingSystem` no longer takes
other `OperatingSystem` objects as an initialization argument. This
is the API break.

Changes to be committed:
	modified:   src/lmn/cx/y22/operating_systems/_operating_systems.py